- Refactored Constants - Updated New Note UI
