+++ app/javascript/src/constants/contact.js
@@ -1,0 +1,28 @@
+export const CONTACTS = [
+  {
+    id: 1,
+    name: "Neeraj Singh",
+    email: "neeraj@email.com",
+    contactnumber: "123456789",
+    department: "Engineering",
+  },
+  {
+    id: 2,
+    name: "Vinay Chandran",
+    email: "vinay@email.com",
+    contactnumber: "123456789",
+    department: "Engineering",
+  },
+  {
+    id: 3,
+    name: "Jagan Kaartik",
+    email: "jagan@email.com",
+    contactnumber: "123456789",
+    department: "Engineering",
+  },
+];
+
+export const SORT_BY_OPTIONS = [
+  { label: "Name", value: "name" },
+  { label: "Age", value: "age" },
+];
