+++ app/javascript/src/constants/note.js
@@ -1,3 +1,5 @@
+import * as yup from "yup";
+
 export const NOTES = [
   {
     id: 1,
@@ -45,26 +47,16 @@ export const ASSIGN_CONTACT_OPTIONS = [
   { label: "Neeraj Singh", value: "Neeraj Singh" },
 ];
 
-export const CONTACTS = [
-  {
-    id: 1,
-    name: "Neeraj Singh",
-    email: "neeraj@email.com",
-    contactnumber: "123456789",
-    department: "Engineering",
-  },
-  {
-    id: 2,
-    name: "Vinay Chandran",
-    email: "vinay@email.com",
-    contactnumber: "123456789",
-    department: "Engineering",
-  },
-  {
-    id: 3,
-    name: "Jagan Kaartik",
-    email: "jagan@email.com",
-    contactnumber: "123456789",
-    department: "Engineering",
-  },
-];
+export const NEW_NOTE_VALIDATION_SCHEMA = yup.object({
+  title: yup.string().required("Title is required"),
+  description: yup.string().required("Description is required"),
+  tags: yup.object().required("Tag is required"),
+  assigncontact: yup.object().required("Assign Contact is required"),
+});
+
+export const NEW_NOTE_INITIAL_VALUES = {
+  title: "",
+  description: "",
+  tags: "",
+  assigncontact: "",
+};
